+++ apps/api-service/src/products/expensetrack/expense/expense.controller.ts
@@ -1,0 +1,98 @@
+import {
+  Controller,
+  Post,
+  Get,
+  Delete,
+  UseGuards,
+  Request,
+  Param,
+  Body,
+  Put,
+  BadRequestException,
+} from "@nestjs/common"
+import { ExpenseService } from "./expense.service"
+import { statusMessages } from "@/shared/constants/status-messages"
+import { AuthGuard, ModRequest } from "@/auth/auth.guard"
+import { CreateExpenseRequestDto } from "./dto/request/create-expense.request.dto"
+
+@Controller("expense")
+export class ExpenseController {
+  constructor(private readonly service: ExpenseService) {}
+
+  @UseGuards(AuthGuard)
+  @Post()
+  async createExpense(
+    @Body() requestBody: CreateExpenseRequestDto,
+    @Request() request: ModRequest
+  ) {
+    try {
+      return await this.service.createExpense(request.user.userId, requestBody)
+    } catch (error) {
+      throw new BadRequestException(
+        error.message || statusMessages.connectionError
+      )
+    }
+  }
+
+  @UseGuards(AuthGuard)
+  @Get()
+  async findMyExpenses(@Request() request: ModRequest) {
+    try {
+      return await this.service.findMyExpenses(request.user.userId)
+    } catch (error) {
+      throw new BadRequestException(
+        error.message || statusMessages.connectionError
+      )
+    }
+  }
+
+  @UseGuards(AuthGuard)
+  @Get("/:expenseId")
+  async findExpenseById(@Request() request: ModRequest, @Param() params: any) {
+    try {
+      return await this.service.findExpenseById(
+        request.user.userId,
+        params.expenseId
+      )
+    } catch (error) {
+      throw new BadRequestException(
+        error.message || statusMessages.connectionError
+      )
+    }
+  }
+
+  @UseGuards(AuthGuard)
+  @Put(":expenseId")
+  async updateExpenseById(
+    @Body() requestBody: CreateExpenseRequestDto,
+    @Param() params: any,
+    @Request() request: ModRequest
+  ) {
+    try {
+      return await this.service.updateExpenseById(
+        request.user.userId,
+        params.expenseId,
+        requestBody
+      )
+    } catch (error) {
+      throw new BadRequestException(
+        error.message || statusMessages.connectionError
+      )
+    }
+  }
+
+  @UseGuards(AuthGuard)
+  @Delete("/:expenseId")
+  async deleteExpense(@Request() request: ModRequest, @Param() params: any) {
+    try {
+      return await this.service.deleteExpense(
+        request.user.userId,
+        params.expenseId
+      )
+    } catch (error) {
+      throw new BadRequestException(
+        error.message || statusMessages.connectionError
+      )
+    }
+  }
+}