+++ apps/api-service/src/products/expensetrack/expense/queries/impl/find-expense-by-user.query.ts
@@ -1,0 +1,3 @@
+export class FindExpensesByUserQuery {
+  constructor(public readonly userId: string) {}
+}